% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_uk.R
\name{aggregate_uk}
\alias{aggregate_uk}
\title{Aggregate UK Data}
\usage{
aggregate_uk(df, level, agg_name, coicops = df$coicop)
}
\arguments{
\item{df}{A data frame containing the data to be aggregated.}

\item{level}{A string indicating the aggregation level.}

\item{agg_name}{The name of the aggregate.}

\item{coicops}{A vector of COICOP codes to be included in the aggregation. Defaults to all COICOP codes in the data frame.}
}
\value{
A data frame with the aggregated data.
}
\description{
This function aggregates UK data based on a specified level and COICOP codes.
}
\examples{
df <- data.frame(date = as.Date('2020-01-01') + 0:365, value = rnorm(366, 100, 10), coicop = rep(letters[1:10], each = 37), m_abb = rep(month.abb, 31))
aggregate_uk(df, "level", "agg_name")
}
