% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_uk.R
\name{chain_uk}
\alias{chain_uk}
\title{Chain UK Data}
\usage{
chain_uk(df, year)
}
\arguments{
\item{df}{A data frame containing the data to be chained.}

\item{year}{The base year for chaining.}
}
\value{
A data frame with the chained data.
}
\description{
This function chains UK data to a base year.
}
\examples{
df <- data.frame(date = as.Date('2020-01-01') + 0:365, laspey = rnorm(366, 100, 10))
chain_uk(df, 2020)
}
