% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib_uk.R
\name{contrib_uk}
\alias{contrib_uk}
\title{Calculate Contribution UK Data}
\usage{
contrib_uk(df, target, value = value, coicops, type = "yoy")
}
\arguments{
\item{df}{A data frame containing the data to be analyzed.}

\item{target}{The target COICOP code for which the contribution is calculated.}

\item{value}{The value column in the data frame.}

\item{coicops}{A vector of COICOP codes to be included in the calculation.}
}
\value{
A data frame with the calculated contributions.
}
\description{
This function calculates the contribution of COICOP codes to a target in UK data.
}
\examples{
df <- data.frame(date = as.Date('2020-01-01') + 0:365, value = rnorm(366, 100, 10), coicop = rep(letters[1:10], each = 37))
contrib_uk(df, "target", value = "value", coicops = letters[1:5])
}
